Create Split a string into array of characters in Python.
In Python, strings are sequences of characters. To manipulate or analyze individual characters, it's often useful to convert the string into an array (or list) of its characters. This process is called splitting a string into a list of characters. This can be done easily using built-in Python features like list() or list comprehensions.